Only enable test framework on feature
