+++ massa-bootstrap/Cargo.toml
@@ -55,4 +55,4 @@ massa_executed_ops = {workspace = true}
 massa_pos_exports = {workspace = true, "features" = ["test-exports"]}
 massa_consensus_exports = {workspace = true, "features" = ["test-exports"]}
 massa_db_worker = {workspace = true, "features" = ["test-exports"]}
-massa_test_framework = {workspace = true}
+massa_test_framework = {workspace = true, "features" = ["test-exports"]}
