+++ massa-test-framework/Cargo.toml
@@ -6,6 +6,9 @@ edition = "2021"
 
 [dependencies]
 massa_hash = { path = "../massa-hash" }
-massa_models = { path = "../massa-models", features = ["test-exports"]}
+massa_models = { path = "../massa-models"}
 massa_signature = { path = "../massa-signature" }
 tracing-subscriber = { workspace = true, features = ["env-filter"]}
+
+[features]
+test-exports = ["massa_models/test-exports"]
